requirements.txt
aiohttp>=3.8.1
orjson>=3.6.7
python-dotenv>=0.20.0
SQLAlchemy>=1.4.32
asyncpg>=0.25.0
uvloop>=0.16.0
pysimdjson>=4.0.3